Contest 2 Task U
